Extract statistics into separate file
