+++ bench/benchmarks_view.js
@@ -7,6 +7,7 @@ versionColor(0); // Skip blue -- too similar to link color.
 
 const formatSample = d3.format(".3r");
 const Axis = require('./lib/axis');
+const {summaryStatistics, regression, kde} = require('./lib/statistics');
 
 class StatisticsPlot extends React.Component {
     constructor(props) {
@@ -15,23 +16,6 @@ class StatisticsPlot extends React.Component {
     }
 
     render() {
-        function kernelDensityEstimator(kernel, ticks) {
-            return function(samples) {
-                if (samples.length === 0) {
-                    return [];
-                }
-                // https://en.wikipedia.org/wiki/Kernel_density_estimation#A_rule-of-thumb_bandwidth_estimator
-                const bandwidth = 1.06 * d3.deviation(samples) * Math.pow(samples.length, -0.2);
-                return ticks.map((x) => {
-                    return [x, d3.mean(samples, (v) => kernel((x - v) / bandwidth)) / bandwidth];
-                });
-            };
-        }
-
-        function kernelEpanechnikov(v) {
-            return Math.abs(v) <= 1 ? 0.75 * (1 - v * v) : 0;
-        }
-
         const margin = {top: 0, right: 20, bottom: 20, left: 0};
         const width = this.state.width - margin.left - margin.right;
         const height = 400 - margin.top - margin.bottom;
@@ -39,8 +23,8 @@ class StatisticsPlot extends React.Component {
 
         const t = d3.scaleLinear()
             .domain([
-                d3.min(this.props.versions.map(v => d3.min(v.samples))),
-                d3.max(this.props.versions.map(v => d3.max(v.samples)))
+                d3.min(this.props.versions.map(v => v.summary.min || Infinity)),
+                d3.max(this.props.versions.map(v => v.summary.max || -Infinity))
             ])
             .range([height, 0])
             .nice();
@@ -51,11 +35,11 @@ class StatisticsPlot extends React.Component {
             .paddingOuter(0.15)
             .paddingInner(0.3);
 
-        const kde = kernelDensityEstimator(kernelEpanechnikov, t.ticks(50));
         const versions = this.props.versions.map(v => ({
             name: v.name,
             samples: v.samples,
-            density: kde(v.samples)
+            summary: v.summary,
+            density: kde(v.samples, t.ticks(50))
         }));
 
         const p = d3.scaleLinear()
@@ -89,17 +73,16 @@ class StatisticsPlot extends React.Component {
                             .domain([0, v.samples.length])
                             .range([0, bandwidth]);
 
-                        const sorted = v.samples.slice().sort(d3.ascending);
-                        const [q1, q2, q3] = [.25, .5, .75].map((d) => d3.quantile(sorted, d));
-                        const mean = d3.mean(sorted);
-
-                        let min = [NaN, Infinity];
-                        let max = [NaN, -Infinity];
-                        for (let i = 0; i < v.samples.length; i++) {
-                            const s = v.samples[i];
-                            if (s < min[1]) min = [i, s];
-                            if (s > max[1]) max = [i, s];
-                        }
+                        const {
+                            mean,
+                            q1,
+                            q2,
+                            q3,
+                            min,
+                            max,
+                            argmin,
+                            argmax
+                        } = v.summary;
 
                         return <g key={i}>
                             <path
@@ -115,7 +98,7 @@ class StatisticsPlot extends React.Component {
                                         fill={color}
                                         cx={scale(i)}
                                         cy={t(d)}
-                                        r={i === min[0] || i === max[0] ? 2 : 1} />
+                                        r={i === argmin || i === argmax ? 2 : 1} />
                                 )}
                                 <line // quartiles
                                     x1={bandwidth / 2}
@@ -152,7 +135,7 @@ class StatisticsPlot extends React.Component {
                                         fontSize={10}
                                         fontFamily='sans-serif'>{formatSample(d)}</text>
                                 )}
-                                {[min, max].map((d, i) =>
+                                {[[argmin, min], [argmax, max]].map((d, i) =>
                                     <text // extent
                                         key={i}
                                         dx={0}
@@ -185,14 +168,6 @@ class StatisticsPlot extends React.Component {
     componentDidMount() {
         this.setState({ width: this.ref.clientWidth });
     }
-}
-
-function regression(samples) {
-    const result = [];
-    for (let i = 0, n = 1; i + n < samples.length; i += n, n++) {
-        result.push([n, samples.slice(i, i + n).reduce(((sum, sample) => sum + sample), 0)]);
-    }
-    return result;
 }
 
 class RegressionPlot extends React.Component {
@@ -291,11 +266,11 @@ class BenchmarkRow extends React.Component {
                                 version.regression.correlation < 0.9 ? '\u2620\uFE0F' :
                                 version.regression.correlation < 0.99 ? '\u26A0\uFE0F' : ''}`)}
                         {this.renderStatistic('Mean',
-                            (version) => `${formatSample(d3.mean(version.samples))} ms`)}
+                            (version) => `${formatSample(version.summary.mean)} ms`)}
                         {this.renderStatistic('Minimum',
-                            (version) => `${formatSample(d3.min(version.samples))} ms`)}
+                            (version) => `${formatSample(version.summary.min)} ms`)}
                         {this.renderStatistic('Deviation',
-                            (version) => `${formatSample(d3.deviation(version.samples))} ms`)}
+                            (version) => `${formatSample(version.summary.deviation)} ms`)}
                     </table>
                     {ended && <StatisticsPlot versions={this.props.versions}/>}
                     {ended && <RegressionPlot versions={this.props.versions}/>}
@@ -350,7 +325,8 @@ for (const name in window.mapboxglBenchmarks) {
             name: ver,
             status: 'waiting',
             logs: [],
-            samples: []
+            samples: [],
+            summary: {}
         };
 
         benchmark.versions.push(version);
@@ -363,7 +339,8 @@ for (const name in window.mapboxglBenchmarks) {
                 .then(samples => {
                     version.status = 'ended';
                     version.samples = samples;
-                    version.regression = leastSquaresRegression(regression(samples));
+                    version.summary = summaryStatistics(samples);
+                    version.regression = regression(samples);
                     update();
                 })
                 .catch(error => {
@@ -387,19 +364,3 @@ function update() {
     );
 }
 
-function leastSquaresRegression(data) {
-    const meanX = d3.sum(data, d => d[0]) / data.length;
-    const meanY = d3.sum(data, d => d[1]) / data.length;
-    const varianceX = d3.variance(data, d => d[0]);
-    const sdX = Math.sqrt(varianceX);
-    const sdY = d3.deviation(data, d => d[1]);
-    const covariance = d3.sum(data, ([x, y]) =>
-        (x - meanX) * (y - meanY)
-    ) / (data.length - 1);
-
-    const correlation = covariance / sdX / sdY;
-    const slope = covariance / varianceX;
-    const intercept = meanY - slope * meanX;
-
-    return { correlation, slope, intercept, data };
-}
